refactor(frontend): rm unused & restructure
