+++ fai-rag-app/fai-frontend/src/App.svelte
@@ -1,36 +1,31 @@
+<script lang="ts" context="module">
+    async function fetchRouteData(path: string, searchParams: URLSearchParams) {
+        const response = await fetch(`/api${path + (searchParams?.toString() ? '?' + searchParams.toString() : '')}`);
+        if (!response.ok) {
+            throw new Error(`Error: ${response.status}`);
+        }
+        return await response.json();
+    }
+</script>
+
 <script lang="ts">
     import {path, searchParams} from 'elegua';
-    import {pageDataStore} from "./lib/store/page";
-    import RenderDynamicComponents from "./lib/components/RenderDynamicComponents.svelte";
     import {writable} from "svelte/store";
+    import type {IComponentDef} from "./lib/types";
+    import RenderPage from "./lib/components/RenderPage.svelte";
+
+    import {pageDataStore} from "./lib/store";
 
     const error = writable<any>(null)
-
-    async function fetchData() {
-        error.set(null);
-        try {
-            const response = await fetch(`/api${$path + ($searchParams?.toString() ? '?' + $searchParams.toString() : '')}`);
-            if (!response.ok) {
-                throw new Error(`Error: ${response.status}`);
-            }
-            const json = await response.json();
-            pageDataStore.set(json);
-        } catch (e: any) {
-            error.set(e.message);
-        }
-    }
 
     $: {
         $path
         $searchParams
-        fetchData()
+        error.set(null)
+        fetchRouteData($path, $searchParams)
+            .then((responseData: IComponentDef[]) => pageDataStore.set(responseData))
+            .catch((e) => error.set(e.message))
     }
-
-
-    $: {
-        $pageDataStore
-    }
-
 </script>
 
 
@@ -39,7 +34,7 @@
         <p class="text-center text-xl text-error">{$error}</p>
     </div>
 {:else if $pageDataStore}
-    <RenderDynamicComponents bind:components={$pageDataStore}/>
+    <RenderPage bind:components={$pageDataStore}/>
 {:else}
     <div class="h-screen flex flex-col items-center justify-center">
         <p class="text-center text-xl text-info">Loading...</p>
