+++ fai-rag-app/fai-frontend/src/lib/async-view/AsyncView.svelte
