+++ fai-rag-app/fai-frontend/src/lib/async-view/async-store.ts
