+++ fai-rag-app/fai-frontend/src/lib/store.ts
@@ -1,6 +1,5 @@
 import {writable} from "svelte/store";
-import type {IComponentDef} from "../types";
+import type {IComponentDef} from "./types";
 
 export const pageDataStore = writable<IComponentDef[]>([])
 
-export default pageDataStore
